Add authentication

+++ app/app.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, HTTPException, File, UploadFile, Form, Depends
-from app.schemas import PostCreate, PostResponse
-from app.db import Post, create_db_and_tables, get_async_session
+from app.schemas import PostCreate, PostResponse, UserRead, UserCreate, UserUpdate
+from app.db import Post, create_db_and_tables, get_async_session, User
 from sqlalchemy.ext.asyncio import AsyncSession
 from contextlib import asynccontextmanager
 from sqlalchemy import select
@@ -10,6 +10,7 @@ import os
 import tempfile
 import shutil
 import uuid
+from app.users import auth_backend, fastapi_users, current_active_user
 
 @asynccontextmanager
 async def lifespan(app: FastAPI):
@@ -18,10 +19,19 @@ async def lifespan(app: FastAPI):
 
 app = FastAPI(lifespan=lifespan)
 
+# We will add in our app all the endpoints that automatically are included by `fastapi_user`. All of them will be located in the prefix.
+app.include_router(fastapi_users.get_auth_router(auth_backend), prefix="/auth/jwt", tags=["auth"])
+app.include_router(fastapi_users.get_register_router(UserRead, UserCreate), prefix="/auth", tags=["auth"])
+app.include_router(fastapi_users.get_reset_password_router(), prefix="/auth", tags=["auth"])
+app.include_router(fastapi_users.get_verify_router(UserRead), prefix="/auth", tags=["auth"])
+app.include_router(fastapi_users.get_users_router(UserRead, UserUpdate), prefix="/users", tags=["users"])
+
+
 @app.post("/upload")
 async def upload_file(
     file: UploadFile = File(...),
     caption: str = Form(...),
+    user: User = Depends(current_active_user),
     session: AsyncSession = Depends(get_async_session) # access de database
 ):
     
@@ -36,12 +46,14 @@ async def upload_file(
             file=open(temp_file_path, 'rb'),
             file_name=file.filename,
             options=UploadFileRequestOptions(
+                use_unique_file_name=True,
                 tags=['backend-upload']
             ),
         )
         
-        if upload_result.error is None:  # success
+        if upload_result.error is None:  # success: code=200
             post = Post(
+                user_id=user.id,
                 caption=caption,
                 url=upload_result.url,
                 file_type="video" if file.content_type.startswith('video/') else 'image',
@@ -61,25 +73,39 @@ async def upload_file(
 
 @app.get("/feed")
 async def get_feed(
+    user: User = Depends(current_active_user),
     session: AsyncSession = Depends(get_async_session) # access de database
 ):
     result = await session.execute(select(Post).order_by(Post.created_at.desc()))
     posts = [row[0] for row in result.all()]
     
+    result = await session.execute(select(User))
+    users = [row[0] for row in result.all()]
+    user_dict = {u.id: u.email for u in users}
+
+    
+    
     posts_data = []
     for post in posts:
         posts_data.append({
             'id': str(post.id),
+            'user_id': str(post.user_id),
             'caption': post.caption,
             'url': post.url,
             'file_type': post.file_type,
             'file_name': post.file_name,
             'created_at': post.created_at.isoformat(),
+            'is_owner': post.user_id == user.id,
+            "email": user_dict.get(post.user_id, "Unknown"),
         })
     return posts_data
 
 @app.delete('/posts/{post_id}')
-async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_session)):
+async def delete_post(
+    post_id: str, 
+    user: User = Depends(current_active_user), 
+    session: AsyncSession = Depends(get_async_session)
+):
     try:
         post_uuid = uuid.UUID(post_id) # goes from string to UUID object
         
@@ -88,6 +114,9 @@ async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_se
         
         if not post:
             raise HTTPException(status_code=404, detail='Post not found')
+        
+        if post.user_id != user.id:
+            raise HTTPException(status_code=403, detail='You are not allowed to delete this post')
         
         await session.delete(post)
         await session.commit()
